feat: Implement steps plot and update steps data type
- Added `PlotSteps.kt` to display steps data in a line chart.
- Changed the `steps` property in `Steps.kt`, `StepsEntity.kt`, and `StepsViewModel.kt` from `Int` to `UInt`.
- Moved `StackedAreaPlotDoubleDataAdapter.kt` from `health.weight` to `utils` package.

+++ composeApp/src/commonMain/kotlin/org/darthacheron/fitbe/health/steps/StepsEntity.kt
@@ -21,7 +21,7 @@ data class StepsEntity(
         return Steps(
             id = id,
             profileId = profileId,
-            steps = steps,
+            steps = steps.toUInt(),
             dateUtc = dateUtc,
         )
     }
@@ -33,7 +33,7 @@ fun toEntity(steps: Steps): StepsEntity {
     return StepsEntity(
         id = steps.id,
         profileId = steps.profileId,
-        steps = steps.steps,
+        steps = steps.steps.toInt(),
         dateUtc = steps.dateUtc,
     )
 }
